Moved timer functions to separate file
