+++ libraries/SuplaDevice/src/supla/network/ENC28J60.h
@@ -20,7 +20,7 @@
 #include <Arduino.h>
 #include <UIPEthernet.h>
 
-#include "../../supla_lib_config.h"
+#include "../supla_lib_config.h"
 #include "network.h"
 
 // TODO: change logs to supla_log
