+++ libraries/SuplaDevice/src/supla/network/esp32_wifi.h
@@ -20,7 +20,7 @@
 #include <Arduino.h>
 #include <WiFi.h>
 
-#include "../../supla_lib_config.h"
+#include "../supla_lib_config.h"
 #include "network.h"
 
 #define MAX_SSID_SIZE          32
